+++ src/commands/add.ts
@@ -1,4 +1,4 @@
-import { SourceControlResourceState, window } from "vscode";
+import { SourceControlResourceState } from "vscode";
 import { Command } from "./command";
 
 export class Add extends Command {
@@ -7,27 +7,12 @@ export class Add extends Command {
   }
 
   public async execute(...resourceStates: SourceControlResourceState[]) {
-    const selection = await this.getResourceStates(resourceStates);
-
-    if (selection.length === 0) {
-      return;
-    }
-
-    const uris = selection.map(resource => resource.resourceUri);
-
-    await this.runByRepository(uris, async (repository, resources) => {
-      if (!repository) {
-        return;
-      }
-
-      const paths = resources.map(resource => resource.fsPath);
-
-      try {
+    await this.executeOnResources(
+      resourceStates,
+      async (repository, paths) => {
         await repository.addFiles(paths);
-      } catch (error) {
-        console.log(error);
-        window.showErrorMessage("Unable to add file");
-      }
-    });
+      },
+      "Unable to add file"
+    );
   }
 }