+++ src/commands/resolve.ts
@@ -13,8 +13,8 @@ export class Resolve extends Command {
     if (selection.length === 0) {
       return;
     }
-    const picks = getConflictPickOptions();
 
+    const picks = getConflictPickOptions();
     const choice = await window.showQuickPick(picks, {
       placeHolder: "Select conflict option"
     });
@@ -23,16 +23,12 @@ export class Resolve extends Command {
       return;
     }
 
-    const uris = selection.map(resource => resource.resourceUri);
-
-    await this.runByRepository(uris, async (repository, resources) => {
-      if (!repository) {
-        return;
-      }
-
-      const files = resources.map(resource => resource.fsPath);
-
-      await repository.resolve(files, choice.label);
-    });
+    await this.executeOnResources(
+      selection,
+      async (repository, paths) => {
+        await repository.resolve(paths, choice.label);
+      },
+      "Unable to resolve conflicts"
+    );
   }
 }