+++ src/commands/revert.ts
@@ -1,4 +1,4 @@
-import { SourceControlResourceState, window } from "vscode";
+import { SourceControlResourceState } from "vscode";
 import { checkAndPromptDepth, confirmRevert } from "../input/revert";
 import { Command } from "./command";
 
@@ -21,19 +21,6 @@ export class Revert extends Command {
       return;
     }
 
-    await this.runByRepository(uris, async (repository, resources) => {
-      if (!repository) {
-        return;
-      }
-
-      const paths = resources.map(resource => resource.fsPath).reverse();
-
-      try {
-        await repository.revert(paths, depth);
-      } catch (error) {
-        console.log(error);
-        window.showErrorMessage("Unable to revert");
-      }
-    });
+    await this.executeRevert(uris, depth);
   }
 }